requirements.txt
jupyter
nbformat
streamlit
pandas
numpy
plotly.express
